feat: split admin rest controllers for different entities
