+++ ldes-server-port-admin-rest/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/admin/rest/config/AdminWebConfig.java
@@ -1,28 +1,30 @@
 package be.vlaanderen.informatievlaanderen.ldes.server.admin.rest.config;
 
 import be.vlaanderen.informatievlaanderen.ldes.server.admin.rest.converters.LdesConfigModelConverter;
-import be.vlaanderen.informatievlaanderen.ldes.server.domain.validation.LdesStreamShaclValidator;
+import be.vlaanderen.informatievlaanderen.ldes.server.domain.validation.LdesConfigShaclValidator;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
-import static be.vlaanderen.informatievlaanderen.ldes.server.domain.ldes.eventstream.config.LdesAdminConstants.SHAPE;
-
 @Configuration
 public class AdminWebConfig {
-	@Bean
-	public LdesConfigModelConverter ldesStreamModelConverter() {
-		return new LdesConfigModelConverter();
-	}
+    @Bean
+    public LdesConfigModelConverter ldesStreamModelConverter() {
+        return new LdesConfigModelConverter();
+    }
 
-	@Bean("streamShaclValidator")
-	public LdesStreamShaclValidator ldesStreamShaclValidator() {
-		// shape needs to be retrieved from the config, needs to be the shacl shape of
-		// the stream
-		return new LdesStreamShaclValidator(SHAPE);
-	}
+    @Bean("streamShaclValidator")
+    public LdesConfigShaclValidator ldesStreamShaclValidator() {
+        return new LdesConfigShaclValidator("streamShaclShape.ttl");
+    }
 
-	@Bean(name = "viewShaclValidator")
-	public LdesStreamShaclValidator ldesViewShaclValidator() {
-		return new LdesStreamShaclValidator(SHAPE);
-	}
+    @Bean(name = "viewShaclValidator")
+    public LdesConfigShaclValidator ldesViewShaclValidator() {
+        //shaclShape for view still needs to be added
+        return new LdesConfigShaclValidator("viewShaclShape.ttl");
+    }
+
+    @Bean(name = "shapeShaclValidator")
+    public LdesConfigShaclValidator ldesShapeShaclValidator() {
+        return new LdesConfigShaclValidator("shapeShaclShape.ttl");
+    }
 }
