+++ ldes-server-port-admin-rest/src/main/java/be/vlaanderen/informatievlaanderen/ldes/server/admin/rest/exceptions/InvalidShaclShapeException.java
@@ -1,0 +1,9 @@
+package be.vlaanderen.informatievlaanderen.ldes.server.admin.rest.exceptions;
+
+import static be.vlaanderen.informatievlaanderen.ldes.server.domain.constants.RdfConstants.RDF_SYNTAX_TYPE;
+
+public class InvalidShaclShapeException extends RuntimeException {
+    public InvalidShaclShapeException(Exception e) {
+        super(e);
+    }
+}
