+++ ldes-server-port-admin-rest/src/test/resources/ldes-1.ttl
@@ -1,6 +1,13 @@
-@prefix tree: <https://w3id.org/tree#> .
 @prefix ldes: <https://w3id.org/ldes#> .
+@prefix dcterms: <http://purl.org/dc/terms/> .
+@prefix tree: <https://w3id.org/tree#>.
+@prefix sh:   <http://www.w3.org/ns/shacl#> .
 
-ldes:collectionName1 a ldes:EventStream ;
-   ldes:versionOfPath "1.1" ;
-   ldes:shape [] .
+
+[] a ldes:EventStream ;
+    ldes:timestampPath dcterms:created ;
+    ldes:versionOfPath dcterms:isVersionOf ;
+    tree:shape [
+        sh:closed "true";
+        a sh:NodeShape ;
+    ] .
